Separate prologue generation and tree.check function.
The prologue being in the same template was a horrible hack and
became unmanageable with imports which are needed for check.

"check" checks the various invariants of a node (the way it's
implemented is a bit of a hack).

Also, remove the generated file when generation has failed for
some reason.