Clean things up a bit.
Rebalance more aggressively on delete. Not that it helps, delete seems
to still unbalance the tree in some rare cases. This is probably
related to me mixing up left and right at some point. I need to clean
up the compares so that cmpval and cmp are always done in the same
direction.

Normally I wouldn't commit something that doesn't fix the bug, but this
makes it better and I'm out of time.